fix(#518): decompose serve.go into serve.go, serve_config.go, serve_plugins.go
Split the 770-line wiring function into three files by responsibility:
- serve.go: cobra command, runServe main flow, buildLogger
- serve_config.go: buildProxyConfig, buildBodySizePortsConfig, buildResiliencePortsConfig, buildEgressPlugin, buildEventLogger
- serve_plugins.go: registerPlugins, buildSecretsPlugin

Pure refactor — no behaviour change.
